+++ pdns/protobuf.cc
@@ -110,22 +110,21 @@ void DNSProtoBufMessage::addTag(const std::string& strValue)
 #endif /* HAVE_PROTOBUF */
 }
 
-void DNSProtoBufMessage::addRR(const std::string& strName, uint32_t uType, uint32_t uClass, uint32_t uTTL, const std::string& strBlob)
+void DNSProtoBufMessage::addRR(const std::string& strName, uint16_t uType, uint16_t uClass, uint32_t uTTL, const std::string& strBlob)
 {
 #ifdef HAVE_PROTOBUF
 
   PBDNSMessage_DNSResponse* response = d_message.mutable_response();
   if (!response)
     return;
-
   PBDNSMessage_DNSResponse_DNSRR* rr = response->add_rrs();
-      if (rr) {
-        rr->set_name(strName.c_str());
-        rr->set_type(uType);
-        rr->set_class_(uClass);
-        rr->set_ttl(uTTL);
-        rr->set_rdata((const uint8_t *) strBlob.c_str(), strBlob.size());
-      }
+  if (rr) {
+    rr->set_name(strName.c_str());
+    rr->set_type(uType);
+    rr->set_class_(uClass);
+    rr->set_ttl(uTTL);
+    rr->set_rdata((const uint8_t *) strBlob.c_str(), strBlob.size());
+  }
 
 #endif /* HAVE_PROTOBUF */
 }
